+++ src/sixlowpan/test/sixlowpan-iphc-test.cc
@@ -142,7 +142,6 @@ SixlowpanIphcImplTest::DoRun()
         rxNode->AddDevice(rxDev);
 
         Ptr<SixLowPanNetDevice> rxSix = CreateObject<SixLowPanNetDevice>();
-        rxSix->SetAttribute("ForceEtherType", BooleanValue(true));
         rxNode->AddDevice(rxSix);
         rxSix->SetNetDevice(rxDev);
 
@@ -165,7 +164,6 @@ SixlowpanIphcImplTest::DoRun()
         txNode->AddDevice(txDev);
 
         Ptr<SixLowPanNetDevice> txSix = CreateObject<SixLowPanNetDevice>();
-        txSix->SetAttribute("ForceEtherType", BooleanValue(true));
         txNode->AddDevice(txSix);
         txSix->SetNetDevice(txDev);